Balance and assistan classes have been created. Class diagram have been updated.
The class balance have been created to let the user have a balance of their operations. The class assistant is going to be the most important class with the information and methods to connect the ui with the model. Also, the classes MoneyLender, User, and Main have some upgrades. Finally, the class diagram have been updated with all this changes.