New exception, and gui improvement.
There's a new exception called "ExistingUserException" and it's thrown when the user tries to create a new user with a used name. Also, the window to add an outlay have been renamed, and there's  a basic version of the window to edit an outlay, finally, there's a little advance in the method to add a new outlay.