Creation and edition of the outlays working perfectly.
The creation and edition of the outlays is working properly. The user can change the value, name, reason or pay date, but he wouldn't  be able to change the type of the outlays. The class checker has two new methods to remove and edit the outlays, the class users also have them.