Search outlay method and window added. Also outlay sort, and some basic GUI improvement.
There's a new fxml file to let the user search outlay(s) using it's name, the methods and window are working perfectly. There's another table column in the list of the outlays so the user can know the creation date of the outlays. Finally, the user can now sort the outlays with the combo box and button in the outlay list window.